repo fork: check if argument is a remote of the current repository
This fixes #2722, in which a Git remote is never added if repo fork
was given an argument.

By default, the remote is only added when no argument is given. Here,
we check if the argument is one of the existing remotes, and consider
that the remote should be added.

THIS IS BRITTLE AND A HACK AND SHOULD NOT BE USED.